combine the code
